# chom/openapi.yaml
openapi: 3.1.0
info:
  title: CHOM SaaS Platform API
  description: |
    Comprehensive REST API for the CHOM multi-tenant SaaS platform for WordPress hosting management.

    ## Features
    - Multi-tenant architecture with organization and tenant isolation
    - WordPress site provisioning and management
    - Automated backups and restore capabilities
    - Team collaboration with role-based access control
    - Real-time metrics and monitoring

    ## Authentication
    All authenticated endpoints require a Bearer token obtained from the `/api/v1/auth/login` or `/api/v1/auth/register` endpoints.
    Include the token in the `Authorization` header:
    ```
    Authorization: Bearer {your-token-here}
    ```

    ## Rate Limiting
    - **Authentication endpoints**: 5 requests per minute
    - **Standard API endpoints**: 60 requests per minute
    - **Sensitive operations** (delete, backup, restore): 10 requests per minute

    ## Pagination
    List endpoints support pagination with the following query parameters:
    - `page` - Page number (default: 1)
    - `per_page` - Items per page (default: 20, max: 100)

    ## Error Handling
    All errors follow a consistent format:
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "Human-readable error message",
        "details": {}
      }
    }
    ```

  version: 1.0.0
  contact:
    name: CHOM API Support
    email: support@chom.example.com
  license:
    name: MIT
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /sites/{id}:
    get:
      tags:
        - Sites
      summary: Get site details
      description: Returns detailed information about a specific site including recent backups.
      operationId: getSite
      parameters:
        - $ref: '#/components/parameters/SiteId'
      responses:
        '200':
          description: Site details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteDetailResponse'
        '404':
          description: Site not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      tags:
        - Sites
      summary: Update site settings
      description: Updates site configuration such as PHP version or custom settings.
      operationId: updateSite
      parameters:
        - $ref: '#/components/parameters/SiteId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSiteRequest'
      responses:
        '200':
          description: Site updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteResponse'
        '404':
          description: Site not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Sites
      summary: Delete a site
      description: Permanently deletes a site and all associated data. This is a destructive operation with rate limiting.
      operationId: deleteSite
      parameters:
        - $ref: '#/components/parameters/SiteId'
      responses:
        '200':
          description: Site deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '404':
          description: Site not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Deletion failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /sites/{id}/enable:
    post:
      tags:
        - Sites
      summary: Enable a site
      description: Enables a disabled site, making it accessible again.
      operationId: enableSite
      parameters:
        - $ref: '#/components/parameters/SiteId'
      responses:
        '200':
          description: Site enabled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteResponse'
        '404':
          description: Site not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /sites/{id}/disable:
    post:
      tags:
        - Sites
      summary: Disable a site
      description: Disables a site, making it temporarily inaccessible.
      operationId: disableSite
      parameters:
        - $ref: '#/components/parameters/SiteId'
      responses:
        '200':
          description: Site disabled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteResponse'
        '404':
          description: Site not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Laravel Sanctum bearer token authentication
  parameters:
    SiteId:
      name: id
      in: path
      required: true
      description: Site UUID
      schema:
        type: string
        format: uuid
      example: 550e8400-e29b-41d4-a716-446655440000
...
  schemas:
...
    UpdateSiteRequest:
      type: object
      properties:
        php_version:
          type: string
          enum: ['8.2', '8.4']
        settings:
          type: object
          additionalProperties: true
    Site:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        domain:
          type: string
          example: example.com
        url:
          type: string
          format: uri
          example: https://example.com
        site_type:
          type: string
          enum: [wordpress, html, laravel]
          example: wordpress
        php_version:
          type: string
          example: '8.2'
        ssl_enabled:
          type: boolean
          example: true
        ssl_expires_at:
          type: string
          format: date-time
          nullable: true
          example: '2024-04-15T10:30:00Z'
        status:
          type: string
          enum: [creating, active, disabled, failed, deleting]
          example: active
        storage_used_mb:
          type: integer
          example: 245
        created_at:
          type: string
          format: date-time
          example: '2024-01-15T10:30:00Z'
        updated_at:
          type: string
          format: date-time
          example: '2024-01-15T10:30:00Z'
        vps:
          type: object
          nullable: true
          properties:
            id:
              type: string
              format: uuid
            hostname:
              type: string
    SiteDetail:
      allOf:
        - $ref: '#/components/schemas/Site'
        - type: object
          properties:
            db_name:
              type: string
              example: wp_example_com
            document_root:
              type: string
              example: /var/www/example.com
            settings:
              type: object
              additionalProperties: true
            recent_backups:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  type:
                    type: string
                  size:
                    type: string
                  created_at:
                    type: string
                    format: date-time
...
    SiteResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Site'
        message:
          type: string
          example: Site is being created.
    SiteDetailResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/SiteDetail'
...
    SuccessMessage:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully.
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: ERROR_CODE
            message:
              type: string
              example: An error occurred.
            details:
              type: object
              additionalProperties: true
...